requirements.txt
Flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10

python-jose[cryptography]==3.3.0
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0